requirements.txt
psycopg2
django-heroku
gunicorn
configparser
pandas
django-localflavor
beautifulsoup4
requests
django-datatables-view